src/requirements.txt
numpy
pandas
faiss-cpu
nltk
sentence_transformers
tqdm  